Updates to valid location QC test, moved land mask data files
The valid location QC test now includes the option to define segments of the radial map to be assigned failure flags.  The option can be used for failing radials in overwater locations that are behind points of land or islands or for locations that not well measured by the radar for any other reason.  The land mask data files used in the mask_over_land method of the Radial class were relocated to hfradarpy/.hfradarpy and this required some changes to setup for pip install.